Added Features - Fixed North South hemispheres
I realised that the Northern and Southern hemispheres display the moon opposite (that the moon waxes from left-to-right in the south, but right-to-left in the North). I have fixed this.

I have also ensured that the Placeholder Deer is actually replaced rather than hidden behind the moon, and that the dimensions of these images are variables for users to edit.

I have taken the 'Flare' off the wallpaper, and made it a few skins. The wallpaper is now starry space and concrete/moon surface triangles.